카드 관련 정보 JSON 직렬화 최적화

+++ src/main/java/com/semoss/agricola/GamePlay/service/AgricolaService.java
@@ -9,8 +9,8 @@ import java.util.List;
 public interface AgricolaService {
     void start(Long gameRoomId);
     void playAction(Long gameRoomId, Long eventId, List<AgricolaActionRequest.ActionFormat> acts);
-    void playExchange(Long gameRoomId, Long improvementId, ResourceType resource, Long count);
-    void playExchange(Long gameRoomId, Long improvementId, AnimalType resource, Long count);
+    void playExchange(Long gameRoomId, Long improvementId, ResourceType resource, int count);
+    void playExchange(Long gameRoomId, Long improvementId, AnimalType resource, int count);
     void finish(Long gameRoomId);
     boolean validatePlayer(Long gameRoomId, Object userId);
 }
